fixed the text box issue, cleaned up code
